chore(runtime): ratchet + robustness refinements from round-2 review
- undeclared_create_field_keys: fail open on a poisoned registry lock instead of
  unwrap() (the action boundary still enforces); accept declared navigation
  properties so a relationship-shaped create field is not wrongly rejected.
- Move the ARN-247 spec_actor and file_initial_writes test modules into sibling
  *_test.rs files so the production files stay under the 500-line readability
  ratchet threshold; no test coverage lost.

Readability ratchet green (PROD_FILES_GT500 + UNWRAP_CI_OK back to baseline).
